+++ pkg/bench/ring_check.go
@@ -9,11 +9,11 @@ import (
 	"time"
 
 	ingester_client "github.com/cortexproject/cortex/pkg/ingester/client"
+	"github.com/cortexproject/cortex/pkg/ring"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/kv/codec"
 	"github.com/grafana/dskit/kv/memberlist"
-	"github.com/grafana/dskit/ring"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/prompb"
 	"github.com/thanos-io/thanos/pkg/discovery/dns"
@@ -72,7 +72,7 @@ func NewRingChecker(id string, instanceName string, cfg RingCheckConfig, workloa
 	cfg.RingConfig.KVStore.MemberlistKV = r.MemberlistKV.GetMemberlistKV
 
 	var err error
-	r.Ring, err = ring.New(cfg.RingConfig, "ingester", ring.IngesterRingKey, logger, reg)
+	r.Ring, err = ring.New(cfg.RingConfig, "ingester", ring.IngesterRingKey, reg)
 	if err != nil {
 		return nil, err
 	}